split shape implementation from constructor
